pyproject.toml
[project]
name = "anikethai"
version = "0.1.0"
description = "A weekend discord bot project"
readme = "README.md"
requires-python = ">=3.10, < 3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.15",
    "aiosignal==1.3.2",
    "async-timeout==4.0.3",
    "attrs==25.3.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "dataclasses-json==0.5.14",
    "discord-py==2.3.0",
    "feedparser==6.0.11",
    "frozenlist==1.5.0",
    "greenlet==3.1.1",
    "idna==3.10",
    "langchain==0.0.202",
    "langchainplus-sdk==0.0.20",
    "marshmallow==3.26.1",
    "multidict==6.3.0",
    "mypy-extensions==1.0.0",
    "numexpr==2.10.2",
    "numpy==1.26.4",
    "openai==0.27.8",
    "openapi-schema-pydantic==1.2.4",
    "packaging==24.2",
    "propcache==0.3.1",
    "pydantic==1.10.21",
    "psycopg2-binary>=2.9",
    "python-dotenv==0.21.1",
    "pytz==2022.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "sgmllib3k==1.0.0",
    "sqlalchemy==2.0.22",
    "tenacity==8.5.0",
    "antlr4-python3-runtime==4.11.1",
    "sympy>=1.13",
    "tqdm==4.67.1",
    "typing-extensions==4.13.0",
    "typing-inspect==0.9.0",
    "urllib3==2.3.0",
    "yarl==1.18.3",
]

[tool.pyright]
venv = ".venv"
venvPath = "."

